requirements.txt
torch
ultralytics
fastapi
uvicorn
Pillow
numpy